Cap iterations of `component_api` fuzzer (#12103)
This timed out over the weekend on OSS-Fuzz and running locally it just
runs quite a lot of iterations of the test case in question. There's no
need to run so much in one fuzz test case so cap the total number of
iterations at a constant to avoid timing out.