Revert "Cranelift: Reassociate long and narrow chains of operations into shallow and wide trees (#7466)" (#12116)
* Revert "Cranelift: Reassociate long and narrow chains of operations into shallow and wide trees (#7466)"

This reverts commit 72534b0f9783570936c83e291e1abdbe7f5e6793.

This mid-end rule was shown to be the root cause of a codegen regression in
https://github.com/bytecodealliance/wasmtime/issues/12106 and we didn't have any
real examples that would improve from this rule, just a theoretical idea of its
benefit. Better to do actual instruction scheduling while lowering and/or during
regalloc than to do hack-y instruction scheduling in the mid-end (where we have
no idea how many registers are available, for example).

* update disas tests