Optimize runtime of `test-artifacts` build script
* Convert core wasms to components in parallel
* Skip the core wasm to component part if the artifacts are up-to-date
  (based on mtime)
* Build the C/C++ programs in parallel.

The goal here is to pick some low-hanging fruit to prevent this from
being such a bottleneck in local development, but there's more that can
be done if necessary (e.g. reading the dep files and calculating that
all manually). Right now for example the longer steps are invoking Cargo
which does nothing and invoking the C/C++ compilers unconditionally, but
solving that makes this more of a "build a build system" script and I
feel like we haven't quite crossed that threshold yet.