filesystem: Paper over fadvise harmless failure on macos
On macos and ios, fadvise is only implemented for WillNeed and
dispatches via the system-interface crate to the F_RDADVISE fcntl.  If
you call WillNeed on an out-of-bounds offset, this fcntl returns
FileTooLarge.  Here we paper over this harmless error code to avoid this
needless platform-dependent nonuniformity.  Should fix wasmtime for
https://github.com/WebAssembly/wasi-testsuite/issues/178.

prtest:full