Fix `#[derive(Lift)]` for enums of 256 cases
This commit fixes a longstanding bug in the implementation of
`#[derive(Lift)]` which was found by the component_api fuzzer recently.
Specifically when an enum or variant had exactly 256 cases the
comparison of the discriminant for being out-of-bounds was done in the
n-bit space of the discriminant rather than a bit-space that can hold
the entire discriminant. The fix here is to compare with a `u32` instead
of a `u8` to ensure that if there are 256 variants it actually compares
against 256 instead of 0 by accident.