Cranelift: Optimize out redundant select + icmp instructions (#12135)
* Implement simple icmp + select optimization

This optimizes the following:

a = select x, k1, k2
b = icmp eq a, k1

to

b = x

We shouldn't trigger this optimization when k1 == k2 because constant
propagation should optimize that case.

* Optimize icmp eq and ne for select-on-icmp (#2)

* Optimize icmp eq and ne for select-on-icmp

When we have a select followed by an icmp and the inner condition
to the select is also an icmp, we can remove the select + icmp and
directly use the inner condition.

We negated the values of the inner_condition via bxor x 1.

* Add trailing newline

* feat: avoid relying on icmp inner_cond

---------

Co-authored-by: school <school@schools-MacBook-Air.local>
Co-authored-by: Edward Wibowo <wibow9770@gmail.com>

* test: add extra test

---------

Co-authored-by: school <school@schools-MacBook-Air.local>
Co-authored-by: Edward Wibowo <wibow9770@gmail.com>