Cranelift: s390x: fix patchable call ABI to not clobber outside clobber-save area.
It turns out that the s390x ABI is special wrt our others: the s390x
System-V ABI provides an area to all callees to save clobbered GPRs.
However this is only large enough for r6-r15, the callee-saves in the
standard ABI.

The implementation implicitly assumed this, and when I adjusted the
definition of clobbered registers for the `patchable` ABI, stating that
r0-r15 are clobbered instead, the code happily generated a
store-multiple (`stmg`) instruction that saved too much data in too
little space, overwriting other bits of the stack.

Also, the clobber-save/restore sequence code only saved the bottom 64
bits of clobbered vector/float registers (which are 128 bits),
implicitly encoding the fact that the SysV ABI specifies only the bottom
half as callee-saved.

This PR implements `patchable` properly on s390x by open-coding the
sequences to save all vector registers and r0-r5 in the explicit
clobber-save area (still using the SysV-defined one for r6-r15).