relax intra-component stream/future read/write rules
Per https://github.com/WebAssembly/component-model/pull/580, we now allow
intra-component reads and writes for all numeric payloads.

While I was implementing this, Alex and I spotted a couple of bugs, which I've
fixed here:

- We can't treat `bool` and `char` payloads as "flat" (i.e. trivially copy-able) since not all bit patterns are valid for those types.
- We weren't enforcing the rules (old or new) for streams with non-"flat" payloads (i.e. the ones that most needed them enforced :facepalm:)

Also, now that https://github.com/WebAssembly/component-model/pull/578 has been
merged, I've updated the `tests/component-model` submodule and reduced the list
of "expect fail" tests for that directory.

Finally, note that the `p3_http_middleware_host_to_host` test is still expected
to trap, but now for a different reason.  Previously, it trapped because it
tried to do an intra-component read/write on a `stream<u8>`.  Now that that's
allowed, it instead traps when doing an intra-component read/write on a
`future<result<option<trailers>, error-code>>`, which is still prohibited.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>