remove `callback_code::POLL` and make `waitable-set.poll` not yield
As of https://github.com/WebAssembly/component-model/pull/578, there is no more
`POLL` callback code, and `waitable-set.poll` should not yield.  Guest
toolchains can emulate the old code with a combination of `YIELD` and
`waitable-set.poll`.

Note that I've removed the now-redundant copy of `trap-if-block-and-sync.wast`
in favor of the upstream version.