Debug: clarify safety around ActivationBacktrace::new. (#12191)
* Debug: clarify safety around ActivationBacktrace::new.

As per [this discussion], we decided that we would handle potential
frame-cursor invalidation unsafety (with a hypothetical future
frame-editing debugger API) by putting `unsafe` on that future
hypothetical API rather than on the cursor construction. With the APIs
available today on the `Store`, there is no way to invalidate the frame
cursor while within its lifetime-bounded scope (with lifetime tied to
the `Store` that is passed into the hostcall creating the cursor), so
the API today should be completely safe. This PR makes it so.

[this discussion]: https://github.com/bytecodealliance/wasmtime/pull/12176#discussion_r2636431864

* store_mut is safe as well.