s390x: Emit 20bit immedate variants of indexed loads
Previously, cranelift did not emit memory operations using
20 bit immedates for reg + reg memory modes. This patch
enables that in a type safe manner by exporting a few
extractors for 20 bit immediates and changing the argument
to the MemArg constructor for reg + reg addressing mode to
accept that 20 bit offset instead of the u8 offset.

Note that the '_bias' suffixed rule is not modified, as it's
primarily used by vector instructions that don't have BXD20
addressing, only BXD12.