`wasmtime-environ`: use `crate::error` instead of `anyhow` (#12203)
* Temporarily re-export `anyhow` as `wasmtime_environ::error`

This is intended to ease the migration to the `wasmtime-internal-error`
crate. Follow up PRs will start moving crates in our workspace over to using
this re-export and not depending on `anyhow` directly themselves. Finally, we
will swap out this re-export for a re-export of the `wasmtime-internal-error`
crate.

* rename existing `error` module to `wasm_error`

* wasmtime-environ: use `crate::error` instead of `anyhow`

Part of migrating to the `wasmtime-internal-error` crate.

Depends on https://github.com/bytecodealliance/wasmtime/pull/12202