Cranelift: pulley: don't use call{1,2,3,4} opcodes with PreserveAll ABI.
These Pulley instructions are an optimization: they incorporate moves
from other registers to x0/x1/x2/x3 as part of the call itself.

This works fine for the default calling convention, but the logic that
absorbs these moves and removes them from the list given to the rest of
the VCode/ABI machinery embeds a tacit assumption: the argument
registers are volatile (caller-saved). In particular, this is because
the call{1,2,3,4} instructions are *not* marked as clobbering
x0/x1/x2/x3 explicitly.

To fix this issue, this PR simply omits the magical folding that happens
in the emission code. An alternative would be to try to add the clobbers
to the special instructions, but actually we don't want these to be just
"clobbers": we want the usual regalloc constraints to take effect and
the values to be represented as continuing valid/live in the registers
after return, and the proper way to do that is to leave the `args` on
the call and let the regalloc operand pass add the fixed-reg operands.

Discovered while trying to land #12183.